requirements.txt
kafka-python
pyspark
pandas
psycopg2